fix: meta tags moved to urls
