+++ admin/src/overview/Urls.jsx
@@ -12,6 +12,10 @@ export default function UrlsOverview( { moduleId } ) {
 					<p>Invaluable for all website owners, managers, and SEO professionals, the URL Monitoring module helps you manage and monitor all the internal and external links throughout your website without the need for much manual input.</p>
 					<p>So, what are some of the most useful features included in the URLsLab URL Monitoring module? Firstly, using the URL Monitoring module you can hide certain URLs, which gives you complete control over what appears on your site. This module also allows you to easily track any changes made to each URL, providing you with valuable data. Furthermore, you can enable the link validation feature to automatically check and report all broken or invalid links. This way you can rectify faulty links before they make a dent in your SEO efforts.</p>
 					<p>Implementing the URL Monitoring module into your daily website operations can lead to higher SEO rankings, increased organic traffic, improved user experience, and enhanced overall website performance.</p>
+					<h3>Meta tags</h3>
+					<p>Meta tags are an invaluable tool for all SEO professionals and website owners alike. Although not visible to users, meta tags help search engines understand your website’s content which allows them to index it better.</p>
+					<p>URLsLab’s Meta Tags Manager module allows you to control and optimize your website’s meta tags. By doing this you effectively dictate how your website is presented in search results. Besides adding and updating meta tags, the Meta Tags Manager module generates open graph meta tags using your pages’ screenshots. This is an invaluable tool when it comes to boosting your social media posts’ visibility. Moreover, this module includes the “add if missing” functionality which ensures all your pages contain appropriate meta tags.</p>
+					<p>It’s no secret that meta tags are the backbone of SEO. Having an effective meta tag optimization strategy is crucial for higher click-through rates and increased organic traffic. With URLsLab’s Meta Tags Manager, you can rest assured that no page will go unoptimized.</p>
 					<h3>Screenshots</h3>
 					<p>Screenshots are a great way to grab an audience's attention and make your content more appealing. With URLsLab's Screenshots module, you can easily insert automatically generated screenshots into your content with just a simple shortcode integration. Not only will you save time, but your content will look more professional.</p>
 					<p>Using the Screenshots module is especially suitable for larger websites with many pages. Instead of manually taking and inserting a screenshot on each page, the Screenshots module will do it for you. You can also choose from various screenshot types, so the result suits your website perfectly.</p>
@@ -101,6 +105,10 @@ export default function UrlsOverview( { moduleId } ) {
 						<p>Yes, you can. URLsLab’s Screenshots module is suitable for websites of all sizes but it can be especially useful for larger sites including many pages.</p>
 						<h4>Can I choose the type of screenshot thayt the module generates?</h4>
 						<p>The Screenshots module gives you four options to choose from: first screen screenshot, full, page screenshot, first screen thumbnail, and full page thumbnail.</p>
+						<h4>How does the Meta Tags Manager module help in improving my website's SEO?</h4>
+						<p>Meta tags help search engines understand and index your website. Optimizing these tags is a reliable way to boost your SEO rankings and attract organic traffic. The Meta Tags module ensures that all your pages contain well-optimized meta tags and, therefore, it directly contributes to your SEO efforts.</p>
+						<h4>Can I use Meta Tags Manager to update existing meta tags on my website?</h4>
+						<p>Yes, you can. You can use the “replace the current value” functionality which allows you to update your meta tags at any time.</p>
 					</section>
 			}
 		</Overview>
