+++ includes/class-urlslab-available-widgets.php
@@ -27,7 +27,6 @@ class Urlslab_Available_Widgets {
 		$urlslab_link_enhancer      = new Urlslab_Widget_Urls();
 		$urlslab_keyword_links      = new Urlslab_Widget_Link_Builder();
 		$urlslab_image_alt_text     = new Urlslab_Widget_Image_Alt_Text();
-		$urlslab_og_meta_tag        = new Urlslab_Widget_Meta_Tag();
 		$urlslab_media_offloader    = new Urlslab_Widget_Media_Offloader();
 		$urlslab_lazy_loading       = new Urlslab_Widget_Lazy_Loading();
 		$urlslab_optimizer          = new Urlslab_Widget_Html_Optimizer();
@@ -46,7 +45,6 @@ class Urlslab_Available_Widgets {
 			$urlslab_serp->get_widget_slug()               => $urlslab_serp,
 			$urlslab_related_resources->get_widget_slug()  => $urlslab_related_resources,
 			$urlslab_faq->get_widget_slug()                => $urlslab_faq,
-			$urlslab_og_meta_tag->get_widget_slug()        => $urlslab_og_meta_tag,
 			$urlslab_media_offloader->get_widget_slug()    => $urlslab_media_offloader,
 			$urlslab_image_alt_text->get_widget_slug()     => $urlslab_image_alt_text,
 			$urlslab_redirects->get_widget_slug()          => $urlslab_redirects,
