+++ includes/widget/class-urlslab-widget-general.php
@@ -258,6 +258,5 @@ class Urlslab_Widget_General extends Urlslab_Widget {
 		( new Urlslab_Api_Permissions() )->register_routes();
 		( new Urlslab_Api_Tasks() )->register_routes();
 		( new Urlslab_Api_Urls() )->register_routes();
-		( new Urlslab_Api_Meta_Tags() )->register_routes();
 	}
 }
